# Conflicts:
#	common/src/main/java/com/gaia3d/ExtensionModuleFrame.java
#	common/src/main/java/com/gaia3d/basic/halfedge/HalfEdgeSurface.java
#	tiler/src/main/java/com/gaia3d/command/mago/Mago3DTilerMain.java
#	tiler/src/main/java/com/gaia3d/process/preprocess/PhotogrammetryMinimizer.java
#	tiler/src/main/java/com/gaia3d/process/tileprocess/tile/PhotogrammetryTiler.java
#	tiler/src/test/java/com/gaia3d/basic/command/UnitTest.java

+++ extension/src/main/java/com/gaia3d/renderer/engine/graph/UniformsMap.java
@@ -24,7 +24,7 @@ public class UniformsMap {
         if (location == null) {
             throw new RuntimeException("Could not find uniform [" + uniformName + "]");
         }
-        return location.intValue();
+        return location;
     }
 
     public void createUniform(String uniformName) {
